Sorting data
Added a filter to be able to sort out unnecessary data from
the text and split the text by \n and \r so the text is
structured in the best way.

The initial data is laying in a single <p> tag on the website
and included a lot of junk text for our purpose.